refactor: split metrics.tsx into modular layers
- Extract type definitions to lib/types/metrics.ts
- Move data fetching logic to lib/services/metrics.ts
- Create utility functions in lib/utils/metrics-formatter.ts
- Split UI into modular components in components/metrics/
- Reduce main file from 426 lines to 8 lines
- Maintain all functionality and Japanese comments
- Improve testability and maintainability
